requirements.txt
pymupdf
pdfplumber
pillow
pdf2image
pytesseract
tqdm
pandas
streamlit
sentence-transformers
faiss-cpu
